refactor(manager): split handlers.go into category-specific files (#33)
Split handlers.go (596 lines) into 5 files:
- handlers-connection.go: OnConnectionStateChange, OnMessage, OnOpen, OnQuit + Remove*
- handlers-filename-events.go: OnFlightLoaded, OnAircraftLoaded, OnFlightPlanActivated + Remove*
- handlers-object-events.go: OnObjectAdded, OnObjectRemoved + Remove*
- handlers-system-events.go: OnCrashed, OnCrashReset, OnSoundEvent, OnView, OnFlightPlanDeactivated + Remove* + Subscribe*
- handlers-simstate.go: OnSimStateChange, OnPause, OnSimRunning + Remove* + Subscribe*

Deleted original handlers.go. No behavior changes — pure file reorganization.
